requirements.txt
amqp==5.0.6
aniso8601==7.0.0
asgiref==3.3.4
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.10.0
billiard==3.6.4.0
cachetools==4.2.4
celery==5.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.0
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cloudinary==1.26.0
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
cryptography==3.4.7
decorator==5.0.9
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.13
django-admin==2.0.1
django-allauth==0.44.0
django-cloudinary-storage==0.3.0
django-cors-headers==3.7.0
django-debug==0.1.1
django-debug-toolbar==3.2.1
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-extensions==3.1.3
django-heroku==0.3.1
django-rest-auth==0.9.5
django-rest-knox==4.1.0
django-rest-swagger==2.2.0
django-restframework==0.0.1
django-seed==0.2.2
django-six==1.0.4
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
dnspython==2.1.0
drf-yasg==1.20.0
email-validator==1.1.3
Faker==8.11.0
flake8==3.9.2
google==3.0.0
google-api-core==2.0.1
google-api-python-client==2.23.0
google-auth==2.2.1
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.1.0
httplib2==0.19.1
idna==2.10
inflection==0.5.1
ipython==7.31.1
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.1
jsonschema==3.2.0
kombu==5.2.2
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mccabe==0.6.1
oauthlib==3.1.1
openapi-codec==1.3.2
packaging==21.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
promise==2.3
prompt-toolkit==3.0.19
protobuf==3.18.0
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
Pygments==2.10.0
PyJWT==2.1.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
rq==1.10.0
rsa==4.7.2
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
Rx==1.6.1
screen==1.0.1
simplejson==3.17.2
singledispatch==3.6.2
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
stripe==2.60.0
swagger-spec-validator==2.7.3
text-unidecode==1.3
traitlets==5.0.5
uritemplate==3.0.1
urllib3==1.26.5
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.3.0
xlwt==1.3.0
